Create alternative to CanValidateField trait
The trait was meant to "wrap" single-source validators which add a
ValidationResult with zero or more violations. The trait is supposed to
add a source name and return only the first violation for cases where we
don't care about each individual violation from a validator. But we
aren't using the trait in this repo and provide it for other code, which
is bad practice, because it's the wrong OOP pattern.

To replace the trait, we add utility methods to ValidationResult to set
the source and get the first violation. Calling code should replace

    $violation = $this->getFieldViolation($result, 'source_name');

with

    $violation = $result->setSource('source_name')->getFirstViolation();

Ticket: https://phabricator.wikimedia.org/T383340